Advancing on debugging test_slam. Added id_factory calls in lmk, rob and sensor constructors. Making sure the measurements have the correct covariance. Still some bugs but the prog is able to run up to TIME 4.

+++ include/rtslam/rtSlam.hpp
@@ -58,7 +58,7 @@ namespace jafar {
 		typedef boost::shared_ptr<Gaussian>						gaussian_ptr_t;
 		typedef boost::shared_ptr<ExtendedKalmanFilterIndirect> ekfInd_ptr_t;
 		
-		typedef kernel::IdFactory<unsigned, kernel::IdCollectorNone> IdFactory; // FIXME maybe we should change for a smarter IdFactory ? eg:
+		typedef kernel::IdFactory<uint64_t, kernel::IdCollectorNone> IdFactory; // FIXME maybe we should change for a smarter IdFactory ? eg:
 		//typedef kernel::IdFactory<unsigned, kernel::IdCollectorList> IdFactory;
 		//typedef kernel::IdFactory<unsigned, kernel::IdCollectorSet> IdFactory;
 	}
